+++ feature/main/presentation/src/main/java/ru/crazerr/feature/main/presentation/main/MainState.kt
@@ -5,21 +5,21 @@ import ru.crazerr.feature.account.domain.api.Account
 data class MainState(
     val accounts: List<Account>,
     val isLoading: Boolean,
-    val currentAmount: Long,
-    val currentIncome: Long,
-    val currentExpenses: Long,
-    val lastMonthExpenses: Long,
-    val lastMonthIncome: Long,
+    val currentAmount: Double,
+    val currentIncome: Double,
+    val currentExpenses: Double,
+    val lastMonthExpenses: Double,
+    val lastMonthIncome: Double,
     val mainCurrencySign: Char,
 )
 
 internal val InitialMainState = MainState(
     accounts = emptyList(),
     isLoading = true,
-    currentAmount = 0L,
-    currentIncome = 0L,
-    currentExpenses = 0L,
-    lastMonthIncome = 0L,
-    lastMonthExpenses = 0L,
+    currentAmount = 0.0,
+    currentIncome = 0.0,
+    currentExpenses = 0.0,
+    lastMonthIncome = 0.0,
+    lastMonthExpenses = 0.0,
     mainCurrencySign = '₽',
 )
