+++ feature/transaction/presentation/src/main/java/ru/crazerr/feature/transaction/presentation/transactionEditorStory/TransactionEditorStoryArgs.kt
@@ -1,5 +1,5 @@
 package ru.crazerr.feature.transaction.presentation.transactionEditorStory
 
 data class TransactionEditorStoryArgs(
-    val transactionId: Int = -1,
+    val transactionId: Long = -1,
 )
